apps/beam/requirements.txt
opencv-python
ffmpeg-python
imutils
numpy
python-telegram-bot
tiktoken
cuid2
faster-whisper
markdownify
requests
filetype